requirements.txt
psutil==5.9.5
GPUtil==1.4.0
pymongo==4.5.0
pandas==2.1.1
numpy==1.26.0
prophet==1.1.3
redis==4.5.5
rq==1.15.1
python-dotenv==1.0.0
streamlit==1.28.0
plotly==5.18.0
pytest==7.4.0
loguru==0.7.2